Rename: Add conflicts between type/namespace declarations
Fixes internal TFS bug #1027506 "Rename: No unresolvable conflict shown
for interface/class name conflict"

Update the ConflictResolver to catch declaration conflicts between types
of different kinds (classes versus interfaces) and between types &
namespaces. When two different kinds of types conflict, the conflict is
shown on the one that's not currently being renamed. When a type
conflicts with a namespace, the conflict is always shown on the type.