Rename: Check for existence of BufferUndoState in UndoTemporaryEdits
Fixes part of internal TFS bug #1167415.

We have reports of the BufferUndoState being unavailable in
UndoTemporaryEdits when inline rename has been dismissed due to an
external workspace change. This change prevents the crash by checking
for the existence of the relevant BufferUndoState while we await more
information on what leads to this state.