Fix threading issue in Generate Type Dialog
Attempting to generate a new file in an ASP.NET 5 web project would
deadlock because we were attempting to get the default namespace from a
worker thread.

The solution is to move the work of computing the default namespace from
ComputeOperationsAsync to GetOptions.  GetOptions is always guaranteed
to execute on the UI thread wheras ComputeOperationsAsync happens from a
worker thread.