Rename HACK_GetTextChangesAsync
Fixes internal TFS bug #1114858

The HACK_GetTextChangesAsync method was originally added when
Document.GetTextChangesAsync was occasionally hanging.
Document.GetTextChangesAsync no longer hangs (in this codepath, at
least) but we still need to keep this method as a way of performing a
document text diff that falls back on using the
IDocumentTextDifferencingService (which in VS uses the editor's text
diffing algorithm) instead of falling back on diffing syntax trees
(which can produce diffs that are too large and disrupt certain
projection scenarios). This change updates the name of the method and
relocates it to below the method where it is consumed.