Add CS0414 to diagnostic analyzer unsupported error codes list
Adds CS0414 (ErrorCode.WRN_UnreferencedFieldAssg / "The field 'X' is
assigned but its value is never used") to the list of error codes that
are not supported in the CSharpCompilerDiagnosticAnalyzer. This allows
CS0414 warnings produced during the build to persist when their relevant
documents are opened in the IDE, which would otherwise cause the
warnings to disappear when the set of live errors is produced.