Fix CA1017 messages
There are really two different FxCop rules that report under the same
ID: one that fires if your assembly is missing a `ComVisible` attribute
entirely, and one that fires if you specify `ComVisible(true)` (in the
latter case, it suggests using `ComVisible(false)` at the assembly
level, and `ComVisible(true)` at the individual class level).

Previously we had just one `DiagnosticDescriptor` representing both,
with a `MessageFormat` of `{0}`. At the point where the `Diagnostic` was
created we did a `string.Format(...)` to create the actual message text,
and then passed it in as the fill-in for the bogus message. This worked
in English but probably would have failed when these messages were
localized into other languages.

The whole point of a `LocalizableString` is to capture enough
information such that the string can be rendered into an arbitrary
locale at a later time; by using `string.Format(...)` we were tying
ourselves to whatever the `ResourceManager` decided was the right locale
to use when the `Diagnostic` was created. This may not be the same
locale that we would pick when actually showing the message to the user.

The fix here is to have two different `DiagnosticDescriptors` that
differ only in the message format.