Get rid of Roslyn.VisualStudio.Setup as a shipping assembly
The RoslynPackage type originally lived in a seperate assembly than
Microsoft.VisualStudio.LanguageServices when it was important for us
to be able to support the core parts of Roslyn being installed atop
the existing Visual Studio 2013 language services. This is no longer
a supported scenario anywhere, so we should just get rid of the
assembly.

Doing this is easy, it's just a matter of moving the files and related
assets to the LanguageServices assembly. A MEF interface is added
to prevent a circularity around triggering a load of stuff in the
SolutionExplorerShim. That could be delt with later by moving the core
of that code back to LanguageServices, but I don't want to touch that
now since it's very sensitive.