Workaround for span mapping in elision buffers
When mapping the span [0,0] from an elision buffer to its subject
buffer, we may receive two spans. If one of the spans is [0,0], we will
prefer the second span. This happens when the user invokes completion in
a workflow editor expression text box before any text has been typed.
Per the editor team, this elision buffer behavior is probably a bug, but
there isn't time to get it fixed for the next release. We'll use this
workaround to avoid crashing.