Fix issue with C# character literals in InitializeComponent() method
In C#, if a WinForms InitializeComponent method() includes character
literal containing a unicode escape for a character that is not valid in
XML, it can cause an XML parser exception to be thrown and the WinForms
designer not to load. To work around this problem, we generate a hidden
cast expression in the IMethodXML to cast the numeric value of the
character to a System.Char.  We do this for any character that isn't a
letter or digit. The designer can process this XML correctly and load
properly.