Add Roslyn to build task so GAC'ing isn't necessary for IDE.
Right now the IDE uses the Csc/Vbc host objects from the build
task by loading the task from the GAC (since the task is only
installed in the MSBuild directory otherwise). However, this breaks
using a different task for ETA tests since the GAC will always
take precedence.

This change adds the build task to the VS setup to include the
task in the IDE installation.