Handle .ruleset deletion/rename properly
Currently, if a .ruleset file is renamed, deleted, or removed, we only
update the `CodeAnalysisRuleSet` property on projects that included the
.ruleset as a project item. This commit institutes better rules:

* If a project- or solution-level .ruleset is renamed, we need to update
every `CodeAnalysisRuleSet` property containing the old name in every C#
and VB project--even if the .ruleset isn't a file in the project.
* If a project- or solution-level .ruleset is deleted, we need to clear
every `CodeAnalysisRuleSet` property containing the name in every C# and
VB project--again, even if the .ruleset isn't a file in the project.
* If a project-level .ruleset is removed, we only clear the
`CodeAnalysisRuleSet` properties for that project.
* If a solution-level .ruleset is removed, we do nothing.

Fixes #1601.