Expand transparent identifier in the (VB) EE
...the same way we expand display classes and state machines.  If we see a
local or parameter with a transparent identifier generated/mangled name,
we should instead include the properties of its (anonymous) type in our
list of bindable symbols.

This is a port of ec4d462d57a7224d07067b18d4c99ef99a368639

There is one notable difference between C# and VB: whereas C# has a
single mangled name for transparent identifiers, VB has four - $VB$It,
$VB$It1, $VB$It2, and $VB$ItAnonymous.  Fortunately, the differences are
not interesting to the EE - they're only to prevent name collisions
(except $VB$ItAnonymous which is used when the value is not an anonymous
type and does not need to be expanded).