Update the VS AnalyzerDependencyChecker
This commit updates the `AnalyzerDependencyChecker` type to reflect the
recent changes to require all analyzers and dependencies to be
specified, and the changes in how we load analyzer assemblies.

Now that we require all analyzers and dependencies to be specified, we
no longer look next to the specified analyzers to optimistically locate
dependencies. Instead, we only consider the list of passed-in analyzers
*plus* the assemblies already loaded into the `AppDomain`. The latter
avoids unhelpful warnings in the event that the analyzer assembly
includes a code fix that take a dependency on a Visual Studio assembly,
which of course can't be included in the list of analyzers. We now
create a warning when we identify a missing dependency.

We now also check analyzers themselves for conflicts (i.e., assemblies
at different paths with the same identity but different contents)
instead of just their dependencies.