Change pipe checksum to SHA256
Change our pipe hashing to SHA256 instead of SHA1.

The use of the SHA1 algorithm in security sensitive situations is
banned.  It is still fine as a simple check summing mechanism for public
data (as we are doing in this case).  But source code scanners can't
understand context and raise warnings whenever it is used at all.  Hence
switching this to SHA256 instead.