Fix deadlock in AliasSymbol.GetAliasTarget
The following line inside GetAliasTarget was causing a deadlock in the
compiler:

bool won = ImmutableInterlocked.InterlockedInitialize(ref
_aliasTargetDiagnostics, newDiagnostics.ToReadOnlyAndFree());

The ToReadOnlyAndFree call caused any lazy diagnostics present in the
DiagnosticsBag to be promptly evaluated.  As a part of evaluation the
diagnostics could call back into other compiler code including
AliasSymbol.GetAliasTarget.  If that happened on the same AliasSymbol
instance the compiler would enter a deadlock as _aliasTarget had already
been set to a non-null value.

The CompareExchange of _aliasTarget is effectively a lock statement
ToReadOnlyAndFree is calling out to untrusted code.  The fix is to stop
doing this by just saving the DiagnosticBag directly vs. calling
ToReadOnlyAndFree.

This resolves DevDiv 1161400