Improve undo stack titles for suggestions applied via light bulb
This change introduces a workspace service that can be used to register the display strings that should be displayed in VS's undo stack for each change applied using light bulb. The string displayed in the undo stack is currently identical to the one displayed in the light bulb menu.

Fixes #273