Improve rename conflicts in nameof
Fixes #1195 "Renaming a member to a name used within a nameof is a
conflict"

When a nameof expression is involved in a rename, conflicts are only
introduced if the nameof previously referenced a non-empty set of
symbols and the post-rename set of referenced symbols contains none of
the previously referenced symbols. This allows overloads to join or exit
the set of references without conflicts, but a conflict is still offered
if the nameof switches from referencing a method to referencing a local
(or similar).