Check consistency after solution load
Check analyzer consistency after solution load.

Right now we only invoke the `AnalyzerDependencyCheckingService` when
explicitly pushing a change in the set of analyzers down to the
workspace. As we batch up all changes during a solution load and wait to
push them to the workspace once at the end, we end up skipping this
check until the next time the user adds or removes an analyzer.

This change causes us to explicitly run the check when solution load
completes.