Check EnvDTE.Configuration.Properties for null
We access the Properties collection of the EnvDTE.Configuration type to
check and update the "CodeAnalysisRuleSet" property. However, sometimes
the Properties collection will be null, leading to a
`NullReferenceException`. For example, the Properties collection always
seems to be null in PowerShell projects.

This change adds the necessary checks to avoid the
`NullReferenceException`.

Related to #2524.