Fix nulls in AnalyzerDependencyChecker
Currently we use a `Task.FromResult(null)` as a sentinel task before a
real dependency check has been queued up. This works fine until we try
to access the results of this task, and things blow up.

Instead, we should create an AnalyzerDependecyResults.Empty singleton,
and create a task from that.

Also, this change catches all exceptions coming out of
`AnalyzerDependencyChecker`, to avoid bringing down VS.