Fix Code Model bug with adding attributes in presence of XML doc comments
Code Model must insert attributes below XML doc comments. Otherwise, this
can change the intent of the user's code.

* In C#, it's bad. If an attribute appears above an XML doc comment for a
member, the attribute is attached to the member but the XML doc comment
will not be emitted.
* In VB, it's worse. If an attribute appears above an XML doc comment for
a member, the attribute isn't actually attached to the member!

The approach taken is pretty simple: if an attribute is being inserted at
the front of the list, just copy the existing leading trivia to it. In
addition, extra care is taken in VB to ensure that the attribute has a
trailing line break.

Unit tests have been added to handle every potential CodeElement that may
encounter this situation. These also uncovered several other bugs,
including lack of support for C# event fields and VB delegate statements.