Extending life time of a sequence scoped local if the local is represented by the sequence and is used as a target of an assignment.
We would need to extend the life time of such local for the same reasons that we do it in a case when sequence is a target of an invocation. - We need to access that local in the outer expression, so it must still be in scope.

I could not create an actual scenario where we would have sequence as a target of an assignment and would need to keep the local until the assignment is done. I still think that is possible in theory and more robust support for this scenario shared with similar support for calls is a good idea.

* refactored the logic that digs out the local so that it could be used in other places.
* made it not optional to specify doNotRelease local to ensure we are not releasing locals by accident
* added comments to the cases where releasing all sequence locals at once is ok.

Fixes #2889