Generate relative paths correctly
The `FilePathUtilities.GetRelativePath(...)` method is used in various
places in the Workspaces and VS layers to generate a relative path from
a starting directory to a target file or directory. However, it has a
bug where it will generate an incorrect path if you need to go up more
than one level from the starting directory to get to the target.

For example, if the starting directory is `C:\Alpha\Beta\Gamma` and the
target is `C:\Alpha\SomeFile.txt`, the correct relative path would be
`..\..\SomeFile.txt`. Instead, we get `..\..\..\SomeFile.txt` (note the
extra `..`).

This commit fixes the issue and adds unit tests for this method.