fixed explicit cache bug in solution crawler
found a bug where we missed caching a project explicitly while doing document analysis.

this is a race so it doesn't happen all the time, but when condition is met, we will re-parse every files in the project to build compilation for every file in the project.

I am not 100% sure since when this bug exists but looks like some recent change caused this to happen more often.

I hope this is the reason for VM increase in some of our perf tests.