Pass correct itemIDs to IVsSymbolicNavigationNotify
Partial fix for internal TFS bug #1169405

We previously always passed Nil for all itemidCodeFile parameters in
IVsSymbolicNavigationNotify to ensure the XAML language service heard
our request (because their current implementation only processes request
for Nil or 0-valued itemids.) This caused Go To Definition and Find All
References to be slower than necessary in non-XAML cases and also
resulted in unwanted behaviors on certain properties that are referenced
in XAML.

Our two options for fixing this were to either:
1. Update our implementation to send the correct item ids for
heuristically "non-generated" files and send Nil for "generated" files
(to match the Dev12 behavior)
2. Always send the actual item ids, even for "generated" files, and
update the XAML language service to do a fast initial check of the
corresponding document to see if they're interested in processing it.

This changeset represents the Roslyn half of option 2. Any builds that
contain this fix but do not contain the corresponding change to the XAML
language service will see no Go To Definition navigation to .xaml files
(e.g. when navigating from a reference to a named Button) and no .xaml
file entries in Find All References.

Unfortunately, this does not remove all heuristics from this algorithm,
because Roslyn must still decide which itemid to pass in the case of a
definition spanning multiple documents (e.g. the MainWindow class
created in default WPF projects). We always prefer the "generated"
documents to give external language services the best opportunity to
participate.