NavBar: Guard against symbol navigation on non-symbol items
Fixes internal TFS bug #1174848

When a Navigation Bar item is chosen, like the NavigationBarSymbolItem
representing a particular member, we present the set of chosen items
(for project, type, and member) as base NavigationBarItems with the text
display properties & tracking spans *copied* out of the real
NavigationBarSymbolItem to remove any display indentation and to make
sure the Navigation Bar does not hold on to old symbols, etc.
When the Navigation Bar is expanded, we present real
NavigationBarSymbolItems again and when they are chosen we perform the
appropriate symbolic navigation.

However, there is a moment after an item is chosen where we have
replaced the real NavigationBarSymbolItem with its NavigationBarItem
copy, but the Navigation Bar is still allowing clicks to the item to
count as navigation (rather than expansion of the collapsed list). When
this happens, we see NavigateToItem called on a base NavigationBarItem
and crash because we believe it will be a NavigationBarSymbolItem.

The base NavigationBarItem was being used for multiple purposes, so this
change makes it abstract and introduces NavigationBarPresentedItem for the
purposes described above and NavigationBarActionlessItem which can be used
for items that legitimately produce no action when selected (used in some
VB event scenarios). The NavigationBarController now rejects any attempts
to select a NavigationBarPresentedItem so that the
INavigationBarItemService implementers will not have to deal with them.

VB was already avoiding this problem by typechecking the
NavigationBarItem, and this change also applies the same type check to
C#.