Fix connection issues in server
This addresses the following issues:

- GitHub #2866: Asserts in the check connection loop.
- DevDiv 1175084: Crashes when server is unable to complete the connection to the named pipe.

These issues are being addressed together because they stem from the same
underlying problem.  The code is not fulfilling two of the main invariants
of connection processing in the server:

- The `Task` representing a client connection should never throw an exception.
Failed connections are instead represented by the data returned from the `Task`.
- A connection should always provide a new timeout value before completing the
compilation process.

The `Connection` type is responsible for the majority of the processing here
and is careful to maintain these invariants.  The `ServerDispatcher` type though
didn't fully account for exceptions.  This meant that errors actually completing
the `NamedPipeClientStream` connection ended up violating both of the core
invariants.

The solution here is two fold:

1. Properly handle the exceptional cases that can come from completing the named
pipe connection.
2. Use a single `Task` to represent a client connection instead of a separate one
for the compilation and keep alive.

close #2866