Reload analyzers with loading errors
If the user adds an analyzer assembly to a project without adding all of
its dependencies we may fail to load its `DiagnosticAnalyzers`. For
example, Microsoft.CodeAnalysis.CSharp.FxCopAnalyzers depends on
Microsoft.CodeAnalysis.FxCopAnalyzers; the analyzers in the former
derive from types in the latter, and so will not load unless we know
about both assemblies.

So if you add Microsoft.CodeAnalysis.CSharp.FxCopAnalyzers to a project
by itself you'll get various warnings in the error list about this, and
the analyzers won't work. However, if you then go and add the missing
dependency (Microsoft.CodeAnalysis.FxCopAnalyzers) the warnings don't go
away and the analyzers still won't work.

The issue is that an `AnalyzerFileReference` only looks for
`DiagnosticAnalyzers` once, and caches the results. So even though we
add the missing dependency and *could* load the types, we don't actually
try.

The fix is in the logic to add a new analyzer assembly in VS. In
addition to pushing a new `AnalyzerFileReference` down to the workspace,
we go through the list of existing analyzers and for each one with load
errors we remove the existing `AnalyzerFileReference` and add a new one.