Support multiple kinds of whitelists
Support multiple kinds of whitelists in the `AnalyzerDependencyChecker`.

Right now the `AnalyzerDependencyChecker` takes a whitelist of
`AssemblyIdentity` as an argument; if an assembly is on this list, we
won't report it as missing when checking that all of an analyzer
assembly's references are present. In practice, VS passes in a list of
all the currently loaded assemblies when running the checker.

However, it is possible that an analyzer assembly has a reference that
VS could satisfy but it simply isn't loaded yet. For example, I've seen
warnings that System.Composition.AttributedModel can't be found, even
though it ships with VS. There's no reliable way to make these warnings
go away.

To fix this, this commit introduces an abstraction,
`IAssemblyWhiteList`, and two implementations. The first,
`AssemblyIdentityWhiteList` simply repackages the functionality we
already have, taking a list of specific `AssemblyIdentity`s. The second,
`AssemblyNamePrefixWhiteList`, whitelists every assembly whose name
starts with a given string.

VS uses these to whitelist all the loaded assemblies along with all
assemblies that start with the name "System". This brings VS in line
with the checks done in VBCSCompiler.exe.