Jenkins cross platform updates
This changes our cross plat build to install the desired version of Mono
on the machine as a part of the CI process.  This lowers the management
overhead of our Mac / Linux machines as we can use any machine in the pool
vs. having reserved, and custom maintained, machines for Roslyn.