Don't restart on failed connection sequence
The compiler should not enter a passive shut down mode in the case where
a client connects and disconnects before it can send any request data.

This fix is a part of our AusGov perf work.  In that scenario we have a
high number of compilations running in parallel.  On medium sized
machines it is possible for the server to quickly become computation
bound and slow to respond to connection requests.  The client will break
off a connection attempt after one second and revert to a straight
command line build.

Before this change we would also restart the server in this case.  That
adds extra overhead because it will be restarted by a later client
request.  With this fix we just keep the server running in this
scenario.

The original intent of passive shut down was to handle Ctrl+C scenarios.
The user hits Ctrl+C and all computation should go away, including the
server.  A failed connection attempt shouldn't be treated as Ctrl+C.