Code Model: When setting the name of a CodeElement, ensure that its members are updated
Code Model tracks nodes with a "node key" concept, which is effectively
the full name of an element coupled with an ordinal. When the name of a
CodeElement is updated, its node key must also be updated since the name
of the underlying node has changed. However, we had failed to update the
node keys of any members of that CodeElement, which means that they can
no longer be located after their parent container's name changes. This
is now fixed.